api/requirements.txt
Django>=5.1.3
djangorestframework>=3.15.2
drf-spectacular>=0.28